Pull CompositionHelper into a module
Fixes #254
